refactor(gateway): split server.rs into focused modules (<1000 lines)
server.rs was 1978 lines; carve it into cohesive modules so every gateway
source file is under 1000 lines:

- session.rs: Copilot session bootstrap (try_import_openclaw_token,
  init_copilot_session).
- listen.rs: the listener/acceptor layer (run_gateway,
  handle_transport_connection).
- chat.rs: chat-frame bookkeeping + context assembly (handle_chat_frame,
  auto_thread_label), delegating the model/tool loop to dispatch.
- thread_handler.rs: the thread/task client-frame handlers.
- admin.rs: config-mutating frames (reload, model switch, agent name,
  working dir).

server.rs (now 907 lines) keeps the per-connection engine: auth, the
recv/select loop, and frame routing to the above. The transport-connection
integration tests stay with the engine. No behavior change.
